+++ changelog.md
@@ -5,6 +5,7 @@ vX.X.X
 - Added icon to colouriser button, closes #115.
 - "Parse ANSI escape sequences" checkbox is now only enabled when decoding mode is ASCII, closes #117.
 - Fixed bug where ANSI escape codes were not disabled when it's checkbox is deselected, closes #119.
+- Added bits/second metrics to the "Stats" sub-tab, closes #120.
 
 v0.6.2
 ------